+++ packages/tsurlfilter/test/rules/rule-factory.test.ts
@@ -2,7 +2,7 @@ import { RuleFactory } from '../../src/rules/rule-factory';
 import { CosmeticRule } from '../../src/rules/cosmetic-rule';
 import { NetworkRule } from '../../src/rules/network-rule';
 import { HostRule } from '../../src/rules/host-rule';
-import { CompatibilityTypes, setConfiguration } from '../../src';
+import { CompatibilityTypes, DEFAULT_RULE_INDEX, setConfiguration } from '../../src';
 
 describe('RuleFactory Builder Test', () => {
     it('works if builder creates correct rules', () => {
@@ -29,7 +29,7 @@ describe('RuleFactory Builder Test', () => {
         expect(rule!.getFilterListId()).toBe(1);
         expect(rule!).toBeInstanceOf(NetworkRule);
 
-        rule = RuleFactory.createRule('127.0.0.1 localhost', 1, false, false, false);
+        rule = RuleFactory.createRule('127.0.0.1 localhost', 1, DEFAULT_RULE_INDEX, false, false, false);
         expect(rule).toBeTruthy();
         expect(rule!.getText()).toBe('127.0.0.1 localhost');
         expect(rule!.getFilterListId()).toBe(1);
@@ -50,13 +50,13 @@ describe('RuleFactory Builder Test', () => {
     it('check host and network rules recognition', () => {
         let rule;
 
-        rule = RuleFactory.createRule('example.org', 1, false, false, true);
+        rule = RuleFactory.createRule('example.org', 1, DEFAULT_RULE_INDEX, false, false, true);
         expect(rule).toBeTruthy();
         expect(rule!.getText()).toBe('example.org');
         expect(rule!.getFilterListId()).toBe(1);
         expect(rule!).toBeInstanceOf(NetworkRule);
 
-        rule = RuleFactory.createRule('example.org', 1, false, false, false);
+        rule = RuleFactory.createRule('example.org', 1, DEFAULT_RULE_INDEX, false, false, false);
         expect(rule).toBeTruthy();
         expect(rule!.getText()).toBe('example.org');
         expect(rule!.getFilterListId()).toBe(1);
@@ -66,13 +66,13 @@ describe('RuleFactory Builder Test', () => {
     it('respects ignore flags', () => {
         let rule;
 
-        rule = RuleFactory.createRule('##.banner', 1, false, true);
+        rule = RuleFactory.createRule('##.banner', 1, DEFAULT_RULE_INDEX, false, true);
         expect(rule).toBeFalsy();
 
-        rule = RuleFactory.createRule('||example.org^', 1, true);
+        rule = RuleFactory.createRule('||example.org^', 1, DEFAULT_RULE_INDEX, true);
         expect(rule).toBeFalsy();
 
-        rule = RuleFactory.createRule('127.0.0.1 localhost', 1, false, false, true);
+        rule = RuleFactory.createRule('127.0.0.1 localhost', 1, DEFAULT_RULE_INDEX, false, false, true);
         expect(rule).toBeFalsy();
     });
 
@@ -86,7 +86,8 @@ describe('RuleFactory Builder Test', () => {
         };
 
         setConfiguration(config);
-        const rule = RuleFactory.createRule('*$denyallow=org|com|example.net', 1, false, true, false);
+        // eslint-disable-next-line max-len
+        const rule = RuleFactory.createRule('*$denyallow=org|com|example.net', 1, DEFAULT_RULE_INDEX, false, true, false);
         expect(rule).toBeTruthy();
     });
 });